+++ cpp/subprojects/boosting/src/boosting/info.cpp
@@ -30,6 +30,11 @@ namespace boosting {
             std::string getTargetArchitecture() const override {
                 return MLRLBOOSTING_TARGET_ARCHITECTURE;
             }
+
+            /**
+             * @see `ILibraryInfo::visitBuildOptions`
+             */
+            void visitBuildOptions(BuildOptionVisitor visitor) const override {}
     };
 
     std::unique_ptr<ILibraryInfo> getLibraryInfo() {