+++ cpp/subprojects/boosting/src/boosting/multi_threading/parallel_rule_refinement_auto.cpp
@@ -19,6 +19,6 @@ namespace boosting {
         } else {
             return getNumAvailableThreads(0);
         }
-    };
+    }
 
 }