+++ cpp/subprojects/common/include/common/omp.hpp
@@ -1,0 +1,10 @@
+/*
+ * @author Michael Rapp (michael.rapp.ml@gmail.com)
+ */
+#pragma once
+
+#include "config.hpp"
+
+#if MULTI_THREADING_SUPPORT_ENABLED
+    #include "omp.h"
+#endif